client-webcam/requirements.txt
opencv-python==4.10.0.84
grpcio==1.66.0
grpcio-tools==1.66.0
protobuf==5.27.3
uuid6==2024.7.10
